# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ORACLE Alpha API
  description: |
    AI-powered trading signal aggregator for Solana with verifiable on-chain performance tracking.
    
    ## Features
    - Multi-source signal aggregation (smart wallets, KOLs, volume, narratives)
    - Weighted scoring algorithm
    - On-chain signal publishing
    - Real-time WebSocket updates
    
    ## Authentication
    Currently no authentication required. Rate limited to 100 req/min.
  version: 1.0.0
  contact:
    name: ShifuSensei
    url: https://github.com/dynamolabs/oracle-alpha
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3900
    description: Local development
  - url: https://oracle-alpha.example.com
    description: Production (example)
...
paths:
...
  /api/signals:
    get:
      tags: [Signals]
      summary: Get trading signals
      parameters:
        - name: minScore
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
          description: Minimum score filter
        - name: maxAge
          in: query
          schema:
            type: integer
          description: Maximum age in minutes
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
          description: Number of results
      responses:
        '200':
          description: List of signals
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  signals:
                    type: array
                    items:
                      $ref: '#/components/schemas/Signal'
  /api/signals/{id}:
    get:
      tags: [Signals]
      summary: Get signal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Signal details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Signal'
        '404':
          description: Signal not found
...
  /api/onchain/signals:
    get:
      tags: [On-Chain]
      summary: Get on-chain published signals
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: On-chain signals
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  signals:
                    type: array
                    items:
                      type: object
...
components:
  schemas:
    Signal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        timestamp:
          type: integer
        token:
          type: string
          description: Solana token address
        symbol:
          type: string
        name:
          type: string
        score:
          type: integer
          minimum: 0
          maximum: 100
        confidence:
          type: integer
        riskLevel:
          type: string
          enum: [LOW, MEDIUM, HIGH, EXTREME]
        sources:
          type: array
          items:
            type: object
            properties:
              source:
                type: string
              weight:
                type: number
              rawScore:
                type: integer
        marketData:
          type: object
          properties:
            mcap:
              type: number
            liquidity:
              type: number
            volume5m:
              type: number
            age:
              type: integer
        analysis:
          type: object
          properties:
            narrative:
              type: array
              items:
                type: string
            strengths:
              type: array
              items:
                type: string
            weaknesses:
              type: array
              items:
                type: string
            recommendation:
              type: string
...